+++ packages/eslint-config/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+-  javascript, javascript-react, light-typescript, light-typescript-react, typescript and typescript-react separate configs
+
 ## [1.0.0]
 
 Empty major bump to reach 1.x.x